Koblet sammen rest-endepunktet med utføring av grensesnittavstemming
