+++ src/main/kotlin/no/nav/familie/oppdrag/rest/AvstemmingController.kt
@@ -6,12 +6,10 @@ import no.nav.familie.oppdrag.repository.OppdragProtokollRepository
 import no.nav.familie.oppdrag.service.AvstemmingService
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.format.annotation.DateTimeFormat
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import java.time.LocalDateTime
 import javax.validation.Valid
 
@@ -21,14 +19,19 @@ import javax.validation.Valid
 class AvstemmingController (@Autowired val oppdragProtokollRepository: OppdragProtokollRepository,
                             @Autowired val avstemmingService: AvstemmingService) {
 
-    @PostMapping(consumes = [MediaType.APPLICATION_JSON_VALUE], path = ["/grensesnittavstemming"])
-    fun utførGrensesnittavstemming(@Valid @RequestBody fomTidspunkt: LocalDateTime, tomTidspunkt: LocalDateTime, fagOmråde: String): ResponseEntity<Ressurs<String>> {
-        val oppdragSomSkalAvstemmes = oppdragProtokollRepository.hentIverksettingerForGrensesnittavstemming(fomTidspunkt, tomTidspunkt, fagOmråde)
+    @PostMapping(path = ["/grensesnittavstemming/{fagsystem}"])
+    fun sendGrensesnittavstemming(@PathVariable("fagsystem") fagsystem: String,
+                                  @RequestParam("fom") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) fom: LocalDateTime,
+                                  @RequestParam("tom") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) tom: LocalDateTime
+    ): ResponseEntity<Ressurs<String>> {
+        OppdragController.LOG.info("Grensesnittavstemming: Kjører for $fagsystem-oppdrag for $fom til $tom")
+        val oppdragSomSkalAvstemmes = oppdragProtokollRepository.hentIverksettingerForGrensesnittavstemming(fom, tom, fagsystem)
 
-        val meldinger = AvstemmingMapper(oppdragSomSkalAvstemmes, fagOmråde).lagAvstemmingsmeldinger()
+        val meldinger = AvstemmingMapper(oppdragSomSkalAvstemmes, fagsystem).lagAvstemmingsmeldinger()
 
         avstemmingService.utførGrensesnittavstemming(meldinger)
-        return ResponseEntity.ok().body(Ressurs.Companion.success("Grensesnittavstemming sendt ok"))
+
+        return ResponseEntity.ok().body(Ressurs.success("Grensesnittavstemming sendt ok"))
     }
 
 }
